# applications/js-gateway/swagger.yaml
openapi: 3.0.0
info:
  title: Portfolio API Gateway
  description: API Gateway for microservices portfolio project
  version: 1.0.0
  contact:
    name: SRE/DevOps Portfolio

servers:
  - url: http://localhost:8082
    description: Local development
  - url: https://api.example.com
    description: Production

paths:
  /health:
    get:
      summary: Health check endpoint
      description: Returns the health status of the API Gateway
      tags:
        - Health
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: healthy
                  service:
                    type: string
                    example: js-gateway
                  timestamp:
                    type: string
                    format: date-time
                  database:
                    type: string
                    example: up

  /v1/auth/register:
    post:
      summary: Register a new user
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - email
                - password
              properties:
                username:
                  type: string
                  example: johndoe
                email:
                  type: string
                  format: email
                  example: john@example.com
                password:
                  type: string
                  format: password
                  example: securepassword123
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User registered successfully
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                      username:
                        type: string
                      email:
                        type: string
        '400':
          description: Bad request
        '409':
          description: User already exists

  /v1/auth/login:
    post:
      summary: Login user
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  example: johndoe
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  user:
                    type: object
        '401':
          description: Invalid credentials

  /api/v1/aggregate:
    get:
      summary: Aggregate data from all services
      description: Calls multiple microservices and aggregates responses
      tags:
        - Aggregation
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Aggregated service data
          content:
            application/json:
              schema:
                type: object
                properties:
                  service:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
                  go_service:
                    type: object
                  python_service:
                    type: object
        '401':
          description: Unauthorized

  /api/v1/process-transaction:
    post:
      summary: Process a transaction
      description: Process a transaction through the go-service
      tags:
        - Transactions
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                customer_id:
                  type: string
                items:
                  type: array
                  items:
                    type: object
                discount_code:
                  type: string
      responses:
        '200':
          description: Transaction processed successfully
        '401':
          description: Unauthorized
        '500':
          description: Service error

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

tags:
  - name: Health
    description: Health check endpoints
  - name: Authentication
    description: User authentication endpoints
  - name: Aggregation
    description: Service aggregation endpoints
  - name: Transactions
    description: Transaction processing endpoints